Give a prefix and bindings for org applications
Related to #1296.

Use `SPC a o` as prefix for org related applications like `org-agenda`
or `org-capture`. The `a` is for "application", and `o` for "org". To
avoid long keystrokes for `agenda`, direct bindings are offered through
`SPC a o <key>`.

Because `org-agenda` is one of the principal application, it has be
mapped to repeating the last letter (`SPC a o o`).

(cherry picked from commit 2b5b2d392244a0f8f1614dc83b98b686b7b14470)